mod files.

+++ app.py
@@ -1,4 +1,5 @@
 """地址轉換為座標"""
+import os
 import time
 from datetime import date
 import pandas as pd
@@ -14,12 +15,58 @@ import requests
 # https://map.tgos.tw/TGOSCloudMap?addr=高雄市前金區五福三路54號1樓
 
 # 定義：從 TGOS 網站取得座標
+# def get_coordinates_from_tgos(addr: str):
+#     """地址轉換為座標函式"""
+#     url = f"https://map.tgos.tw/TGOSCloudMap?addr={addr}"
+#     options = Options()
+#     options.add_argument("--headless=new")
+
+
+#     try:
+#         browser = webdriver.Chrome(options=options)
+#     except WebDriverException as e:
+#         print(f"無法啟動瀏覽器：{e}")
+#         return None, None
+
+#     try:
+#         browser.get(url)
+#         wait = WebDriverWait(browser, 10)
+
+        # location_div = wait.until(
+        #     EC.presence_of_element_located((By.CSS_SELECTOR, "body > div:nth-child(45) > div > div > div.modal-body > div"))
+        #     )
+
+        # # 判斷是否包含指定字串
+        # if "搜尋不到任何地址" in location_div.text:
+        #     print(f"地址「{addr}」查無座標資料。")
+        #     return 0, 0
+
+#         location_div = wait.until(
+#             EC.presence_of_element_located((By.CSS_SELECTOR, "div.fw-normal.text-secondary"))
+#             )
+#         try:
+#             span_text = location_div.find_element(By.TAG_NAME, "span").text
+#             full_text = location_div.text
+#             coords_text = full_text.replace(span_text, "").strip()
+#             lat, lng = coords_text.split(",")
+#             return lat.strip(), lng.strip()
+#         except (ValueError, NoSuchElementException) as e:
+#             print(f"解析座標失敗：{e}")
+#             return None, None
+#     except TimeoutException:
+#         print(f"地址「{addr}」載入逾時，無法取得座標。")
+#         return 0, 0
+#     except Exception as e:
+#         print(f"地址「{addr}」取得座標時發生錯誤：{e}")
+#         return 0, 0
+#     finally:
+#         browser.quit()
+
 def get_coordinates_from_tgos(addr: str):
     """地址轉換為座標函式"""
     url = f"https://map.tgos.tw/TGOSCloudMap?addr={addr}"
     options = Options()
     options.add_argument("headless")
-
 
     try:
         browser = webdriver.Chrome(options=options)
@@ -40,24 +87,24 @@ def get_coordinates_from_tgos(addr: str):
             print(f"地址「{addr}」查無座標資料。")
             return 0, 0
 
-        location_div = wait.until(
-            EC.presence_of_element_located((By.CSS_SELECTOR, "div.fw-normal.text-secondary"))
-            )
-        try:
-            span_text = location_div.find_element(By.TAG_NAME, "span").text
-            full_text = location_div.text
-            coords_text = full_text.replace(span_text, "").strip()
-            lat, lng = coords_text.split(",")
-            return lat.strip(), lng.strip()
-        except (ValueError, NoSuchElementException) as e:
-            print(f"解析座標失敗：{e}")
-            return None, None
+        # 等待 li 元素出現
+        li_tag = wait.until(EC.presence_of_element_located((By.CLASS_NAME, "complex-locate-item")))
+
+        lat = li_tag.get_attribute("data-lat")
+        lng = li_tag.get_attribute("data-lng")
+
+        if not lat or not lng:
+            print(f"地址「{addr}」查無座標資料。")
+            return 0, 0
+
+        return lat.strip(), lng.strip()
+
     except TimeoutException:
         print(f"地址「{addr}」載入逾時，無法取得座標。")
-        return None, None
+        return 0, 0
     except Exception as e:
         print(f"地址「{addr}」取得座標時發生錯誤：{e}")
-        return None, None
+        return 0, 0
     finally:
         browser.quit()
 
@@ -67,10 +114,12 @@ today = date.today().strftime("%Y%m%d")
 FILE_NAME = r'data/全國5大超商資料集.csv'
 NEW_FILE_NAME = f'data/全國5大超商資料集_{today}.csv'
 
-INDEX_URL = r"https://data.gcis.nat.gov.tw/od/file?oid=C054F05C-0A6B-428C-B388-288BDB0618E4"
-response = requests.get(INDEX_URL, timeout=10)
-with open(NEW_FILE_NAME, "wb") as f:
-    f.write(response.content)
+if not os.path.exists(NEW_FILE_NAME):
+    INDEX_URL = r"https://data.gcis.nat.gov.tw/od/file?oid=C054F05C-0A6B-428C-B388-288BDB0618E4"
+    response = requests.get(INDEX_URL, timeout=10)
+    with open(NEW_FILE_NAME, "wb") as f:
+        f.write(response.content)
+    print('索引檔下載完成。')
 
 # 讀取 CSV 檔案
 df = pd.read_csv(FILE_NAME)
